chore(definitelyTyped): add possibility to run specific types to ease debug
